+++ eslint-config/test/data/bad.js
@@ -2,7 +2,7 @@
 * @file Test file with errors
 *
 * @author Ty-Lucas Kelley <ty@greylocklabs.com> (https://greylocklabs.com)
-* @copyright Copyright (c) 2017 Greylock Labs. See LICENSE file for details.
+* @copyright Copyright (c) 2017-2019 Greylock Labs. See LICENSE file for details.
 */
 
 import crypto from 'crypto'