+++ eslint-config/test/index.test.js
@@ -2,7 +2,7 @@
  * @file Tests for ESLint rules
  *
  * @author Ty-Lucas Kelley <ty@greylocklabs.com> (https://greylocklabs.com)
- * @copyright Copyright (c) 2017-2018 Greylock Labs. See LICENSE file for details.
+ * @copyright Copyright (c) 2017-2019 Greylock Labs. See LICENSE file for details.
  */
 
 import shell from 'shelljs';